[NOTE] we squash this PR to avoid the existance of some dead experimental code for smaller software size we can always retrieve it in `guided_gen` branch. This PR is an edit based on https://github.com/ise-uiuc/nnsmith/pull/23

* fix requirement

* Guide reshape.

* torch manual seed

* pool mode: execute z3 in standalone process

* option yes to 'Report folder already exists'

* keep stdout stderr in report folder

* support specify fix_nodes vs max_nodes.

* Debugging utils.

* guided_gen bug fix

* Record hostname and environmental options.

* fix limnfloats bugs

* Group reshape

* debug reshape

* group reshape v2

* Improve group reshape.

* Ablation.

* move group constraint to reshape

* bitvec slicing

* exp: limnf separately

* limit nfloats individually; increase mem limit

* update merge op

* restore cons_reshape

* mark todo

* add bits option; fix check_arith_ref

* fix flops

* option to control limnf

* debug bv

* fix forward backward prob

* adapt insert_input to ph; init ph dtype set to float32

* autopep

* Fix linear shape_fn.

* Finer-grained control on random ranks.

* Fix scalar reshape; Fix ShapeVar eq.

* Minor fix.

* Better log message.

* Mark incomplete modeling for furture notice.

* Fix pooling.

* Slice op

* Adapt guidance to placeholder.

* Fix CLI

* Add slice guidance; Integrate guidance with backward gen.

* Minor fix.

* Flexible bcast deduct rank; constrain flops.

* Improve slice.

* Improve slice.

* Default not limit nfloats in z3.

* Default to customized forkserver (aka 'pool') mode.

* Switch default reshape config.

* Don't delete .profraw.

* Support flush llvm profile.

* Update readme.

* Integrate guidance for new ops.

* Fix linear.

* Better debug info.

* Record model in fuzz.

* Record placeholder constraints.

* Rename group reshape.

* Integrate guidance to backward input constraint.

* Smaller.

* Random lower bound instead of binning

* Update gitignore.

* Constrain flops by default.

* Fix typo.

* Default to 30s.

* Experiment scripts.

* minor

* Trim guidance configs.

* No reshape guidance.

* Mitigate floating point exception.

* Improve scripts.

* Revert and fix zero division error.

* Add ORTDebug backend.

* Best effort guidance to avoid graph diversity.

* Avoid hang.

* Integrate nfloat limit.

* Hybrid gen.

* Constrain input shape after graph finalized.

* Optimize; also constrain n_float completely in random gen.

* Default to limit nfloat.

* opt: Reduce bin size to 5.

* opt: no fork; one-sided bin to avoid search space loss; seperate bin constraint.

* Fix ort nthreads.

* Tweak.

* Add option to not run backend.

* Fallback to fork to be consistent with previous setting.

* Option to not count model gen time.

* Fix rank.

* Fix merge.

* Fix summary.

* Add missing reshape op.

* WIP: integrate guidance to variable-sized list.

* Fix guidance: remove extra_constraint.

* Normalize prob for similar ops.

* Complete pad guidance.

* Recommend forkserver.

* Fix slice.

* Fix hybrid mode.

* Debug.

* Consistent to main.

* Skip tvm dtype test.

* Minor: add onnx check at graph_gen.py

* Integrate guidance and hybrid.

* Param diversity analyzer.

* Bar plot.

* Fuzz cuda.

* input shape analysis

* Fix path.

* Lint and unit test.

* Remove unused code.

* Remove check_sat.

* Support tar.

* merge: dce

* clean: block list

* update rdm

+++ experiments/cov_eval.py
@@ -57,6 +57,8 @@ if __name__ == '__main__':
                         60 * 4, help='max time in seconds for coverage evaluation')
     parser.add_argument('--llvm-version', type=str, default='',
                         help='version of llvm during coverage stuff. must align w/ tvm.')
+    parser.add_argument('-y', action='store_true')
+    parser.add_argument('--keep_raw', action='store_true')
     args = parser.parse_args()
 
     # Set global seed
@@ -83,7 +85,7 @@ if __name__ == '__main__':
         args.dp = 1000
 
     if not args.resume:
-        mkdir(args.report_folder)
+        mkdir(args.report_folder, yes=args.y)
         # FORMAT:
         #   batch time (gen time + eval time)
         #   seed
@@ -150,7 +152,7 @@ if __name__ == '__main__':
             # for memcov
             memcov_name = f'{i}.memcov'
             memcov_path = os.path.join(args.report_folder, memcov_name)
-            
+
             seed = random.getrandbits(32)
 
             model_batch = []
@@ -177,7 +179,8 @@ if __name__ == '__main__':
             ]
             if args.memcov:
                 arguments += ['--memcov', memcov_path]
-            p = subprocess.Popen(arguments, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=copied_env)
+            p = subprocess.Popen(
+                arguments, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=copied_env)
             _, errs = p.communicate()
             errs = errs.decode()
             exit_code = p.returncode
@@ -193,7 +196,7 @@ if __name__ == '__main__':
             stderr_file.flush()
 
             if '$ORT.SKIP$' in errs:  # all models are unsupported by ort. skip it.
-                if os.path.exists(profraw_path):
+                if os.path.exists(profraw_path) and not args.keep_raw:
                     os.remove(profraw_path)
                 continue
 
@@ -221,20 +224,22 @@ if __name__ == '__main__':
                         if args.sum:
                             os.system(
                                 f'{llvm_cov} report -instr-profile={profdata_path} {lib_expr} > {sum_path}')
-                        assert 0 == os.system(f'lz4 {lcov_path}')
-                        os.remove(profraw_path)
-                        os.remove(profdata_path)
+                        assert 0 == os.system(
+                            f'lz4 {lcov_path} {lcov_path}.lz4')
+                        if not args.keep_raw:
+                            os.remove(profraw_path)
+                            os.remove(profdata_path)
                 else:
                     print(f'{profraw_path} does not exist...', file=sys.stderr)
 
             if os.path.exists(lcov_path + '.lz4'):
                 record(btime + lagged_time, len(model_batch) +
-                    lagged_n_model, seed, cov_name=lcov_name)
+                       lagged_n_model, seed, cov_name=lcov_name)
                 lagged_time = 0
                 lagged_n_model = 0
             elif os.path.exists(memcov_path + '.pkl'):
                 record(btime + lagged_time, len(model_batch) +
-                    lagged_n_model, seed, cov_name=memcov_name)
+                       lagged_n_model, seed, cov_name=memcov_name)
                 lagged_time = 0
                 lagged_n_model = 0
             else:
